fix(app-server-transport): discard stale pairing responses
