+++ codex-rs/app-server-transport/src/transport/remote_control/pairing.rs
@@ -20,6 +20,7 @@ pub(super) struct RemoteControlPairingClient {
     environment_id: String,
     expires_at: OffsetDateTime,
     auth_change_revision: u64,
+    generation: u64,
 }
 
 impl RemoteControlPairingClient {
@@ -30,6 +31,7 @@ impl RemoteControlPairingClient {
         environment_id: String,
         expires_at: OffsetDateTime,
         auth_change_revision: u64,
+        generation: u64,
     ) -> Self {
         Self {
             pairing_url: remote_control_target.pair_url.clone(),
@@ -38,11 +40,16 @@ impl RemoteControlPairingClient {
             environment_id,
             expires_at,
             auth_change_revision,
+            generation,
         }
     }
 
     pub(super) fn matches_auth_change_revision(&self, auth_change_revision: u64) -> bool {
         self.auth_change_revision == auth_change_revision
+    }
+
+    pub(super) fn generation(&self) -> u64 {
+        self.generation
     }
 
     pub(super) async fn start(
