+++ codex-rs/app-server-transport/src/transport/remote_control/tests.rs
@@ -65,6 +65,7 @@ const TEST_REMOTE_CONTROL_URL: &str = "http://127.0.0.1:1/backend-api/wham/remot
 const TEST_REMOTE_CONTROL_SERVER_TOKEN: &str = "Remote Control Token";
 const TEST_REFRESHED_REMOTE_CONTROL_SERVER_TOKEN: &str = "Refreshed Remote Control Token";
 const TEST_REMOTE_CONTROL_SERVER_TOKEN_EXPIRES_AT: &str = "2999-01-01T00:00:00Z";
+const TEST_HTTP_ACCEPT_TIMEOUT: Duration = Duration::from_secs(10);
 
 fn remote_control_auth_manager() -> Arc<AuthManager> {
     auth_manager_from_auth(CodexAuth::create_dummy_chatgpt_auth_for_testing())
@@ -151,12 +152,16 @@ fn remote_control_handle_with_pairing_client(
         "env_test".to_string(),
         OffsetDateTime::from_unix_timestamp(33_336_362_096).expect("future timestamp should parse"),
         /*auth_change_revision*/ 0,
+        /*generation*/ 0,
     ))));
     RemoteControlHandle {
         enabled_tx: Arc::new(enabled_tx),
         status_tx: Arc::new(status_tx),
         state_db_available: true,
-        pairing_client,
+        pairing: PairingClientState {
+            client: pairing_client,
+            generation: Arc::new(std::sync::atomic::AtomicU64::new(0)),
+        },
         auth_change_rx: Arc::new(StdMutex::new(auth_change_rx)),
     }
 }
@@ -928,221 +933,8 @@ async fn remote_control_handle_enable_disable_stops_and_restarts_connections() {
     let _ = remote_task.await;
 }
 
-#[tokio::test]
-async fn remote_control_handle_disable_clears_stale_pairing_client() {
-    let remote_handle = remote_control_handle_with_pairing_client(
-        TEST_REMOTE_CONTROL_URL,
-        watch::channel(/*init*/ 0u64).1,
-    );
-
-    assert_eq!(
-        remote_handle.disable(),
-        RemoteControlStatusChangedNotification {
-            status: RemoteControlConnectionStatus::Disabled,
-            server_name: test_server_name(),
-            installation_id: TEST_INSTALLATION_ID.to_string(),
-            environment_id: None,
-        }
-    );
-    remote_handle.enable().expect("enable should succeed");
-    assert_eq!(
-        remote_handle
-            .start_pairing(RemoteControlPairingStartParams { manual_code: false })
-            .await
-            .expect_err("re-enabled remote control should wait for refreshed pairing auth")
-            .to_string(),
-        "remote control pairing is unavailable until enrollment completes"
-    );
-}
-
-#[tokio::test]
-async fn remote_control_handle_rejects_pairing_client_after_auth_change() {
-    let (auth_change_tx, auth_change_rx) = watch::channel(/*init*/ 0u64);
-    let remote_handle =
-        remote_control_handle_with_pairing_client(TEST_REMOTE_CONTROL_URL, auth_change_rx);
-    auth_change_tx.send_modify(|revision| *revision += 1);
-
-    assert_eq!(
-        remote_handle
-            .start_pairing(RemoteControlPairingStartParams::default())
-            .await
-            .expect_err("pairing should wait for current-account enrollment")
-            .to_string(),
-        "remote control pairing is unavailable until enrollment completes"
-    );
-}
-
-#[tokio::test]
-async fn remote_control_handle_discards_pairing_response_after_auth_change() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("listener should bind");
-    let remote_control_url = remote_control_url_for_listener(&listener);
-    let codex_home = TempDir::new().expect("temp dir should create");
-    save_auth(
-        codex_home.path(),
-        &remote_control_auth_dot_json(Some("account_id")),
-        AuthCredentialsStoreMode::File,
-    )
-    .expect("initial auth should save");
-    let auth_manager = AuthManager::shared(
-        codex_home.path().to_path_buf(),
-        /*enable_codex_api_key_env*/ false,
-        AuthCredentialsStoreMode::File,
-        /*chatgpt_base_url*/ None,
-    )
-    .await;
-    let remote_handle = remote_control_handle_with_pairing_client(
-        &remote_control_url,
-        auth_manager.auth_change_receiver(),
-    );
-    let pairing_task = tokio::spawn({
-        let remote_handle = remote_handle.clone();
-        async move {
-            remote_handle
-                .start_pairing(RemoteControlPairingStartParams::default())
-                .await
-        }
-    });
-
-    let pairing_request = accept_http_request(&listener).await;
-    assert_eq!(
-        pairing_request.request_line,
-        "POST /backend-api/wham/remote/control/server/pair HTTP/1.1"
-    );
-    assert_eq!(
-        pairing_request.headers.get("authorization"),
-        Some(&format!("Bearer {TEST_REMOTE_CONTROL_SERVER_TOKEN}"))
-    );
-    assert_eq!(
-        serde_json::from_str::<serde_json::Value>(&pairing_request.body)
-            .expect("pairing request body should deserialize"),
-        json!({ "manual_code": false })
-    );
-
-    save_auth(
-        codex_home.path(),
-        &remote_control_auth_dot_json(Some("next_account_id")),
-        AuthCredentialsStoreMode::File,
-    )
-    .expect("next auth should save");
-    auth_manager.reload().await;
-    respond_with_json(
-        pairing_request.stream,
-        json!({
-            "pairing_code": "stale-pairing-code",
-            "manual_pairing_code": "ABCD-EFGH",
-            "server_id": "srv_e_test",
-            "environment_id": "env_test",
-            "expires_at": "3026-05-22T12:34:56Z",
-        }),
-    )
-    .await;
-
-    assert_eq!(
-        pairing_task
-            .await
-            .expect("pairing task should join")
-            .expect_err("stale pairing response should be discarded")
-            .to_string(),
-        "remote control pairing is unavailable until enrollment completes"
-    );
-}
-
-#[tokio::test]
-async fn remote_control_handle_clears_pairing_client_after_auth_change() {
-    let listener = TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("listener should bind");
-    let remote_control_url = remote_control_url_for_listener(&listener);
-    let codex_home = TempDir::new().expect("temp dir should create");
-    save_auth(
-        codex_home.path(),
-        &remote_control_auth_dot_json(Some("account_id")),
-        AuthCredentialsStoreMode::File,
-    )
-    .expect("initial auth should save");
-    let auth_manager = AuthManager::shared(
-        codex_home.path().to_path_buf(),
-        /*enable_codex_api_key_env*/ false,
-        AuthCredentialsStoreMode::File,
-        /*chatgpt_base_url*/ None,
-    )
-    .await;
-    let (transport_event_tx, _transport_event_rx) =
-        mpsc::channel::<TransportEvent>(CHANNEL_CAPACITY);
-    let shutdown_token = CancellationToken::new();
-    let (remote_task, remote_handle) = start_remote_control(
-        RemoteControlStartConfig {
-            remote_control_url,
-            installation_id: TEST_INSTALLATION_ID.to_string(),
-        },
-        Some(remote_control_state_runtime(&codex_home).await),
-        auth_manager.clone(),
-        transport_event_tx,
-        shutdown_token.clone(),
-        /*app_server_client_name_rx*/ None,
-        /*initial_enabled*/ true,
-    )
-    .await
-    .expect("remote control should start");
-
-    let enroll_request = accept_http_request(&listener).await;
-    respond_with_json(
-        enroll_request.stream,
-        remote_control_server_token_response(
-            "srv_e_initial",
-            "env_initial",
-            TEST_REMOTE_CONTROL_SERVER_TOKEN,
-        ),
-    )
-    .await;
-    let mut first_websocket = accept_remote_control_connection(&listener).await;
-
-    save_auth(
-        codex_home.path(),
-        &remote_control_auth_dot_json(Some("next_account_id")),
-        AuthCredentialsStoreMode::File,
-    )
-    .expect("next auth should save");
-    auth_manager.reload().await;
-    expect_remote_control_connection_closed(
-        &mut first_websocket,
-        "auth change should close the stale websocket",
-    )
-    .await;
-    assert_eq!(
-        remote_handle
-            .start_pairing(RemoteControlPairingStartParams::default())
-            .await
-            .expect_err("pairing should wait for current-account enrollment")
-            .to_string(),
-        "remote control pairing is unavailable until enrollment completes"
-    );
-
-    let enroll_request = accept_http_request(&listener).await;
-    assert_eq!(
-        enroll_request.request_line,
-        "POST /backend-api/wham/remote/control/server/enroll HTTP/1.1"
-    );
-    respond_with_json(
-        enroll_request.stream,
-        remote_control_server_token_response(
-            "srv_e_next",
-            "env_next",
-            TEST_REFRESHED_REMOTE_CONTROL_SERVER_TOKEN,
-        ),
-    )
-    .await;
-    let mut second_websocket = accept_remote_control_connection(&listener).await;
-    second_websocket
-        .close(None)
-        .await
-        .expect("second websocket should close");
-
-    shutdown_token.cancel();
-    let _ = remote_task.await;
-}
+#[path = "pairing_integration_tests.rs"]
+mod pairing_integration_tests;
 
 #[tokio::test]
 async fn remote_control_transport_clears_outgoing_buffer_when_backend_acks() {
@@ -2231,7 +2023,7 @@ async fn expect_remote_control_connection_closed(
 }
 
 async fn accept_http_request(listener: &TcpListener) -> CapturedHttpRequest {
-    let (stream, _) = timeout(Duration::from_secs(5), listener.accept())
+    let (stream, _) = timeout(TEST_HTTP_ACCEPT_TIMEOUT, listener.accept())
         .await
         .expect("HTTP request should arrive in time")
         .expect("listener accept should succeed");
